pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "crewai-demo"
version = "0.1.0"
description = "A demonstration of CrewAI framework capabilities through an autonomous trip planning assistant"
readme = "README.md"
requires-python = ">=3.10"
license = {text = ""}
authors = [
    {name = "norooziosos@gmail.com"}
]
dependencies = [
    "crewai>=0.118.0",
    "crewai-tools>=0.42.2",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest",
]

[tool.setuptools]
packages = ["crews", "tools", "utils"]

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120 